Added comments, changed benchmark result set to iterators

+++ include/random.h
@@ -1,8 +1,9 @@
 #ifndef GAL_RAND_H
 #define GAL_RAND_H
+#include <cstddef>
 
-unsigned gal_rand(unsigned upTo);
-unsigned gal_rand(unsigned from, unsigned upTo);
+size_t gal_rand(size_t upTo);
+size_t gal_rand(size_t from, size_t upTo);
 
 double gal_rand(double upTo);
 double gal_rand(double from, double upTo);
